Added Verify Command
Added the verify command, and tested it definitely working. Also added isVerified field to database so we can track a verified user across multiple servers.

TODO:
Add a check when the bot starts, or when a user joins the server, to see if the user is verified in the database, if yes, add the discord role, and vice versa.